% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID2Name.R
\name{ID2Name}
\alias{ID2Name}
\title{Function to convert term IDs to names in data frame}
\usage{
ID2Name(x, col.name = "term_name")
}
\arguments{
\item{x}{data.frame with a column containg GO IDs}

\item{col.name}{character indicating the column name containg the GO IDs to be converted}
}
\value{
data.frame with new column containing GO term names from GO IDs
}
\description{
Function to convert term IDs to names in data frame
}
